Mobt 157 nbhood refactor move plugins (#1691)
* Moved all neighbourhoood plugins into one module.

* Updated checksums after acceptance test removed in previous PR.

* Updated test coverage of GeneratePercentilesFromANeighbourhood

* Removed promotion of single point percentile and realization dimensions to dim coords
