+++ improver/nowcasting/lightning.py
@@ -39,7 +39,7 @@ from iris.exceptions import ConstraintMismatchError
 
 from improver import PostProcessingPlugin
 from improver.metadata.probabilistic import find_threshold_coordinate
-from improver.nbhood.square_kernel import NeighbourhoodProcessing
+from improver.nbhood.nbhood import NeighbourhoodProcessing
 from improver.utilities.cube_checker import check_cube_coordinates
 from improver.utilities.rescale import apply_double_scaling, rescale
 from improver.utilities.temporal import (
