+++ improver_tests/acceptance/SHA256SUMS
@@ -339,7 +339,6 @@ b3f74d83ab680c9800c0c767831de1f170b121d8ded6cdc37db8d8fe5e9b0070  ./nbhood/basic
 2867efbd7cba8592a98d5eba8caa1f13fca20d85fb0730ddfca1ca1a98a1ed57  ./nbhood/mask/input_masked.nc
 5d27b767c86671b176ea6180e6608b22202aaad1564f658f2b3999a7b46bd36a  ./nbhood/mask/kgo_external_masked.nc
 0bab68194852947cb01e36431a0ef4319e3d5385f690308d46fb2cb1f67bcc54  ./nbhood/mask/kgo_masked.nc
-59cd097e12f346cb103fffe7d2208068be262aa33a9f2f74dca1a14276016cc4  ./nbhood/mask/kgo_re_mask.nc
 51436871de0a3a89a49395979d7d3f33eac39532381496244d49d0a27dfa1a0d  ./nbhood/mask/mask.nc
 76c1276afe044fbfac517ffb59cc600ab360b754d448277fa7a581a9710d5e4d  ./nbhood/percentile/input_circular_percentile.nc
 9f55225d6f8888c4a0c807df74dfc34ce1969eb0b37ac814bd55d3d79675d95b  ./nbhood/percentile/kgo_circular_percentile.nc
