+++ improver_tests/psychrometric_calculations/precip_phase_probability/test_PrecipPhaseProbability.py
@@ -37,7 +37,7 @@ import numpy as np
 from cf_units import Unit
 from iris.tests import IrisTest
 
-from improver.nbhood.circular_kernel import GeneratePercentilesFromANeighbourhood
+from improver.nbhood.nbhood import GeneratePercentilesFromANeighbourhood
 from improver.psychrometric_calculations.precip_phase_probability import (
     PrecipPhaseProbability,
 )
